requirements.txt
Django==1.9.1
Pillow==3.1.0
djangorestframework==3.3.0
dj-database-url==0.3.0
django-ckeditor==5.0.3
psycopg2==2.6.1
